Factor out asciidoc.ts
