+++ src/test/code-sample.test.ts
@@ -1,43 +1,7 @@
-import {extractSamples, stripSource, applyPrefixes} from '../code-sample';
+import {stripSource, applyPrefixes} from '../code-sample';
 import {dedent} from '../utils';
-
-const ASCII_DOC1 = `
-
-For instance, this is a valid TypeScript program:
-
-[[greet-ts]]
-[source,ts]
-----
-function greet(who: string) {
-  console.log('Hello', who);
-}
-----
-
-But when you run this through a program like \`node\` that expects JavaScript, you'll get an error:
-
-----
-function greet(who: string) {
-                  ^
-
-SyntaxError: Unexpected token :
-----
-
-`;
-
-const ASCIIDOC_NO_ID = `
-blah blah
-
-[source,js]
-----
-console.log('Hello', 'node');
-----
-
-[source,ts]
-----
-console.log('Hello', 'TS');
-----
-
-`;
+import { extractSamples } from '../asciidoc';
+import { baseSample } from './common';
 
 const ASCIIDOC_PREPEND = `
 
@@ -82,84 +46,32 @@ const c: ABC = 'c';
 ----
 `;
 
-const ASCIIDOC_SKIP = `
-
-// verifier:skip (this isn't really TypeScript)
-[source,ts]
-----
-keyof A&B = (keyof A) | (keyof B)
-----
-`;
-
 describe('code-sample', () => {
-  const baseSample = {
-    tsOptions: {},
-    nodeModules: [],
-    isTSX: false,
-    checkJS: false,
-    sectionId: null,
-    sectionHeader: null,
-    sourceFile: 'source.asciidoc',
-  };
-
-  const baseExtract = {
-    ...baseSample,
-    prefixes: [],
-  };
-
-  describe('extractSamples', () => {
-    test('basic', () => {
-      expect(extractSamples(ASCII_DOC1, 'doc1', 'source.asciidoc')).toEqual([
+  test('prepend directive', () => {
+    expect(applyPrefixes(extractSamples(ASCIIDOC_PREPEND, 'prepend', 'source.asciidoc'))).toEqual(
+      [
         {
-          ...baseExtract,
+          ...baseSample,
           language: 'ts',
-          id: 'greet-ts',
+          id: 'prefix',
+          content: `type AB = 'a' | 'b';`,
+        },
+        {
+          ...baseSample,
+          language: 'ts',
+          id: 'combined',
           content: dedent`
-            function greet(who: string) {
-              console.log('Hello', who);
-            }`,
+          type AB = 'a' | 'b';
+          const a: AB = 'a';`,
         },
-      ]);
-    });
-
-    test('no ID', () => {
-      // Only the TypeScript sample gets extracted (may want to revisit this).
-      expect(extractSamples(ASCIIDOC_NO_ID, 'noid', 'source.asciidoc')).toEqual([
         {
-          ...baseExtract,
+          ...baseSample,
           language: 'ts',
-          id: 'noid-10',
-          content: `console.log('Hello', 'TS');`,
+          id: 'final',
+          content: `const a: AB = 'a';`,
         },
-      ]);
-    });
-
-    test('prepend directive', () => {
-      expect(applyPrefixes(extractSamples(ASCIIDOC_PREPEND, 'prepend', 'source.asciidoc'))).toEqual(
-        [
-          {
-            ...baseSample,
-            language: 'ts',
-            id: 'prefix',
-            content: `type AB = 'a' | 'b';`,
-          },
-          {
-            ...baseSample,
-            language: 'ts',
-            id: 'combined',
-            content: dedent`
-            type AB = 'a' | 'b';
-            const a: AB = 'a';`,
-          },
-          {
-            ...baseSample,
-            language: 'ts',
-            id: 'final',
-            content: `const a: AB = 'a';`,
-          },
-        ],
-      );
-    });
+      ],
+    );
   });
 
   test('multiple prepend directives', () => {
@@ -283,194 +195,6 @@ describe('code-sample', () => {
         import _ from 'lodash';
         const p = {name: 'Bob'};
         const {name} = p;`,
-      },
-    ]);
-  });
-
-  test('skip directive', () => {
-    expect(extractSamples(ASCIIDOC_SKIP, 'skip', 'source.asciidoc')).toEqual([]);
-  });
-
-  test('tsconfig directive', () => {
-    expect(
-      extractSamples(
-        dedent`
-    // verifier:tsconfig:noImplicitAny=false
-    // verifier:tsconfig:strictNullChecks=false
-    [[implicit-any]]
-    [source,ts]
-    ----
-    const x: number = null;
-    ----
-
-    // verifier:tsconfig:noImplicitAny=true
-    // verifier:tsconfig:strictNullChecks=true
-    [[strict]]
-    [source,ts]
-    ----
-    const x: number = null;
-    ----
-
-    `,
-        'tsconfig',
-        'source.asciidoc',
-      ),
-    ).toEqual([
-      {
-        ...baseExtract,
-        language: 'ts',
-        id: 'implicit-any',
-        content: `const x: number = null;`,
-        tsOptions: {
-          noImplicitAny: false,
-          strictNullChecks: false,
-        },
-      },
-      {
-        ...baseExtract,
-        language: 'ts',
-        id: 'strict',
-        content: `const x: number = null;`,
-        tsOptions: {
-          noImplicitAny: true,
-          strictNullChecks: true,
-        },
-      },
-    ]);
-  });
-
-  test('prepend-with-id', () => {
-    expect(
-      extractSamples(
-        ASCIIDOC_PREPEND +
-          '\n' +
-          dedent`
-    // verifier:reset
-
-    // verifier:prepend-id-to-following:prefix
-    // verifier:prepend-id-to-following:combined
-    [[prepended-with-ids]]
-    [source,ts]
-    ----
-    console.log(a);
-    ----
-    `,
-        'prepend-with-id',
-        'source.asciidoc',
-      ).slice(-1),
-    ).toEqual([
-      {
-        ...baseExtract,
-        language: 'ts',
-        id: 'prepended-with-ids',
-        content: `console.log(a);`,
-        prefixes: [{id: 'prefix'}, {id: 'combined'}],
-      },
-    ]);
-  });
-
-  test('next-is-tsx', () => {
-    expect(
-      extractSamples(
-        dedent`
-    // verifier:next-is-tsx
-    [[tsx-example]]
-    [source,ts]
-    ----
-    console.log(a);
-    ----
-    `,
-        'tsx-example',
-        'source.asciidoc',
-      ).slice(-1),
-    ).toEqual([
-      {
-        ...baseExtract,
-        language: 'ts',
-        id: 'tsx-example',
-        content: `console.log(a);`,
-        isTSX: true,
-      },
-    ]);
-  });
-
-  test('header resets', () => {
-    expect(
-      extractSamples(
-        dedent`
-      == Chapter 1
-      // verifier:prepend-to-following
-      [source,ts]
-      ----
-      const x = 12;
-      ----
-
-      [[chapter-2]]
-      == Chapter 2
-      [source,ts]
-      ----
-      const x = 12;
-      ----
-    `,
-        'header-reset',
-        'source.asciidoc',
-      ),
-    ).toEqual([
-      {
-        ...baseExtract,
-        id: 'header-reset-4',
-        sectionHeader: 'Chapter 1',
-        language: 'ts',
-        content: `const x = 12;`,
-      },
-      {
-        ...baseExtract,
-        id: 'header-reset-11',
-        sectionId: 'chapter-2',
-        sectionHeader: 'Chapter 2',
-        language: 'ts',
-        content: `const x = 12;`,
-      },
-    ]);
-
-    expect(
-      extractSamples(
-        dedent`
-      == Chapter 1
-      // verifier:prepend-to-following
-      [source,ts]
-      ----
-      const x = 12;
-      ----
-
-      ==== Chapter 2
-      [source,ts]
-      ----
-      const x = 12;
-      ----
-    `,
-        'header-reset',
-        'source.asciidoc',
-      ),
-    ).toEqual([
-      {
-        ...baseExtract,
-        id: 'header-reset-4',
-        sectionHeader: 'Chapter 1',
-        language: 'ts',
-        content: `const x = 12;`,
-      },
-      {
-        ...baseExtract,
-        id: 'header-reset-10',
-        sectionHeader: 'Chapter 1',
-        language: 'ts',
-        content: `const x = 12;`,
-        prefixes: [
-          {
-            id: 'header-reset-4',
-          },
-        ],
       },
     ]);
   });
